requirements.txt
crewai
python-dotenv
streamlit
folium
streamlit-folium
